Refactored python script building, fixed according to PR comments

+++ scripts/test.sh
@@ -27,26 +27,30 @@ function iterate_dir(){
     network="$1"
     echo "Starting tests on $network"
     for test_case in "$test_dir"/*; do
-        total=$((total + 1))
-        test_name=$(basename $test_case)
-        echo "Test $total) $test_name"
+        if [[ "$network" == "devnet" ]] && [[ "$test_case" == "wallet-test" ]]; then
+            echo "Skipping devnet test for wallet-test"
+        else
+            total=$((total + 1))
+            test_name=$(basename $test_case)
+            echo "Test $total) $test_name"
 
-        config_file_path="$test_case/$CONFIG_FILE_NAME"
-        if [ ! -f "$config_file_path" ]; then
-            echo "No config file provided!"
-            continue
+            config_file_path="$test_case/$CONFIG_FILE_NAME"
+            if [ ! -f "$config_file_path" ]; then
+                echo "No config file provided!"
+                continue
+            fi
+
+            # replace the dummy config (CONFIG_FILE_NAME) with the one used by this test
+            /bin/cp "$config_file_path" "$CONFIG_FILE_NAME"
+
+            NETWORK="$network" "$test_case/check.sh" && success=$((success + 1)) || echo "Test failed!"
+
+            rm -rf starknet-artifacts
+            git checkout --force
+            git clean -fd
+            echo "----------------------------------------------"
+            echo
         fi
-
-        # replace the dummy config (CONFIG_FILE_NAME) with the one used by this test
-        /bin/cp "$config_file_path" "$CONFIG_FILE_NAME"
-
-        NETWORK="$network" "$test_case/check.sh" && success=$((success + 1)) || echo "Test failed!"
-
-        rm -rf starknet-artifacts
-        git checkout --force
-        git clean -fd
-        echo "----------------------------------------------"
-        echo
     done
     echo "Finished tests on $network"
 }
